requirements.txt
blinker==1.9.0
click==8.1.7
flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
numpy==2.2.0
scipy==1.14.1
threadpoolctl==3.5.0
Werkzeug==3.1.3
langchain_community
langchain-huggingface
langchain-chroma
langchain_mistralai
spacy